+++ domain/src/main/java/com/micahnyabuto/domain/usecases/GetExpensesUsecase.kt
@@ -7,7 +7,7 @@ import com.micahnyabuto.domain.repository.ExpensesRepository
 class GetExpensesUseCase(
     private val expensesRepository: ExpensesRepository
 ) {
-    suspend operator fun invoke(userId: Int): Result<List<Expense>> {
+    suspend operator fun invoke(userId: Long): Result<List<Expense>> {
         return expensesRepository.getAllExpenses(userId)
     }
