+++ domain/src/main/java/com/micahnyabuto/domain/usecases/GetUserDataUseCase.kt
@@ -7,7 +7,7 @@ import com.micahnyabuto.domain.repository.ExpensesRepository
 class GetUserDataUseCase(
     private val expensesRepository: ExpensesRepository
 ) {
-    suspend operator fun invoke(userId: Int): Result<User> {
+    suspend operator fun invoke(userId: Long): Result<User> {
         return expensesRepository.getUserData(userId)
     }
 }
